resolved slot repetition and changed project setup
