--- autoJoin.js
@@ -1,13 +1,0 @@
-setInterval(function() {
-    let micButton = document.getElementsByClassName("U26fgb JRY2Pb mUbCce kpROve uJNmj HNeRed QmxbVb")[0];
-    micButton.click();
-    console.debug("Clicked mic button");
-
-    let camButton = document.getElementsByClassName("U26fgb JRY2Pb mUbCce kpROve uJNmj HNeRed QmxbVb")[0];
-    camButton.click();
-    console.debug("Clicked cam button");
-
-    let joinButton = document.getElementsByClassName("uArJ5e UQuaGc Y5sE8d uyXBBb xKiqt")[0];
-    joinButton.click();
-    console.debug("Clicked join button");
-}, 10000);
