--- manifest.json
@@ -1,26 +1,0 @@
-{
-    "name": "Google Meet Auto Join",
-    "version": "1.0",
-    "description": "Most of us are facing issues in managing google meet links for our online lectures. Here's a solution! Add your links in this extension and never worry about forgetting about lectures. The extension will open the link at the given time and auto join the meeting and this repeats every week",
-    "content_scripts" : [
-        {
-            "matches": ["<all_urls>"],
-            "js": ["popup.js"]
-        }
-    ],
-    "background": {
-        "scripts": ["background.js"],
-        "persistent": true
-    },
-    "permissions": [
-        "storage",
-        "alarms",
-        "tabs",
-        "http://*/",
-        "https://*/"
-    ],
-    "browser_action": {
-        "default_popup": "popup.html"
-    },
-    "manifest_version": 2
-}
