+++ src/pages/Popup/scripts/alarm.js
@@ -1,7 +1,7 @@
 function sendMessage(type, key) {
   let message = {};
-  message["type"] = type;
-  message["key"] = key;
+  message['type'] = type;
+  message['key'] = key;
   chrome.runtime.sendMessage(message, () => {});
 }
 
@@ -27,12 +27,12 @@ async function removeDataFromStorage(key) {
 
 //called from popup to removes alarms associated to a particular course id
 //alarmsKeyList : Array of alarms ids
-async function RemoveAlarms(alarmsKeyList){
+async function RemoveAlarms(alarmsKeyList) {
   console.log(alarmsKeyList);
-  for (let i = 0; i < alarmsKeyList.length; i++){
+  for (let i = 0; i < alarmsKeyList.length; i++) {
     var key = alarmsKeyList[i];
     await removeDataFromStorage(key);
-    sendMessage("REMOVE_ALARM", key);
+    sendMessage('REMOVE_ALARM', key);
   }
 }
 
@@ -43,12 +43,12 @@ class Alarm {
   constructor(course, startdata) {
     this.generateID = function () {
       var dt = new Date().getTime();
-      var uuid = "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(
+      var uuid = 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(
         /[xy]/g,
         function (c) {
           var r = (dt + Math.random() * 16) % 16 | 0;
           dt = Math.floor(dt / 16);
-          return (c == "x" ? r : (r & 0x3) | 0x8).toString(16);
+          return (c == 'x' ? r : (r & 0x3) | 0x8).toString(16);
         }
       );
       return uuid;
@@ -67,14 +67,26 @@ function get_nearestDate(day, time) {
   var alarmHours = parseInt(time.split(':')[0]);
   var alarmMinutes = parseInt(time.split(':')[1]) - 1;
   var currentDate = new Date();
-  var days = ["Sunday", "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday"];
+  var days = [
+    'Sunday',
+    'Monday',
+    'Tuesday',
+    'Wednesday',
+    'Thursday',
+    'Friday',
+    'Saturday',
+  ];
   var alarmDay = days.indexOf(day);
   var currentDay = currentDate.getDay();
   var diff = alarmDay - currentDay;
   console.log(diff);
-  if(diff < 0) diff = 7 + diff;
-  if(diff ==  0){
-    if(currentDate.getHours() > alarmHours && currentDate.getMinutes() > alarmMinutes) diff = 7;
+  if (diff < 0) diff = 7 + diff;
+  if (diff == 0) {
+    if (
+      currentDate.getHours() > alarmHours &&
+      currentDate.getMinutes() > alarmMinutes
+    )
+      diff = 7;
   }
   var numberOfDaysToAdd = diff;
   currentDate.setDate(currentDate.getDate() + numberOfDaysToAdd);
@@ -103,17 +115,17 @@ function get_nearestDate(day, time) {
     1: {start: "14:30", end: "14:55"} 
     }
 */
-function AddAlarm_click(course, dates){
+function AddAlarm_click(course, dates) {
   for (var day in dates) {
-    for (var i = 0; i < dates[day].length; i++){
-      course['time'] = dates[day][i]
+    for (var i = 0; i < dates[day].length; i++) {
+      course['time'] = dates[day][i];
       var start_time = dates[day][i]['start'];
       var start_date = get_nearestDate(day, start_time);
       var alarm_data = new Alarm(course, start_date);
       setDataIntoStorage(alarm_data.id, alarm_data);
-      sendMessage("ADD_ALARM", alarm_data.id);
+      sendMessage('ADD_ALARM', alarm_data.id);
     }
   }
 }
 
-export { AddAlarm_click, getAllDataFromStorage, RemoveAlarms }
+export { AddAlarm_click, getAllDataFromStorage, RemoveAlarms };
