--- style.css
@@ -1,106 +1,0 @@
-.wrapper {
-    margin: 0 auto;
-    padding-right: 10px;
-    padding-left:  10px;
-    width: 90vw;
-}
-
-.wrapper input {
-    border: none;
-    border-bottom: 1px black solid;
-    padding: 10px;
-    margin: 10px auto;
-    width: 100%;
-    box-sizing: border-box;
-}
-
-.wrapper .input-with-icon {
-    padding-left: 30px;
-}
-
-.wrapper input:focus{
-    outline: none;
-}
-
-.wrapper button {
-    height: 30px;
-    width: 90%;
-    outline: none;
-    border-radius: 20px;
-    border: 2px solid #08645A;
-    color: black;
-    cursor: pointer;
-    background: white;
-    transition: background 0.1s ease-in-out;
-    margin: 10px;
-}
-
-.wrapper button:hover {
-    background: #08645A;
-    color: white;
-    border: none;
-}
-
-.popup-title {
-    color: #08645A;
-    margin: 20px;
-    text-align: center;
-}
-
-.input-label, .data-list {
-    color: #08645A;
-    margin: 10px;
-}
-
-.wrapper ul {
-    display: block;
-    list-style-type: disc;
-    margin-block-start: 1em;
-    margin-block-end: 1em;
-    margin-inline-start: 0px;
-    margin-inline-end: 0px;
-    padding-inline-start: 0px;
-}
-
-.wrapper li {
-    margin: 10px auto;
-    font-size: medium;
-}
-
-.wrapper .alarm_list {
-    margin: 0px 10px;
-}
-
-
-.ChromeExtension::-webkit-scrollbar {
-    width: 10px;
-}
-
-.ChromeExtension::-webkit-scrollbar-track {
-    box-shadow: inset 0 0 5px grey; 
-    border-radius: 10px;
-}
-
-.ChromeExtension::-webkit-scrollbar-thumb {
-    background: #08645A; 
-    border-radius: 10px;
-}
-
-.ChromeExtension::-webkit-scrollbar-thumb:hover {
-    background: #053d37; 
-}
-
-.right-inner-addon {
-    position: relative;
-    
-    width: 90vw;
-}
-
-.right-inner-addon svg {
-    position: absolute;
-    left: -10px;
-    padding: 10px 12px;
-    pointer-events: none;
-    fill: #08645A;
-    top: 8px;
-}
